# Conflicts:
#	data/shaders/opengl/attributes.glsl
#	src/Body.cpp
#	src/Makefile.am
#	src/ModelViewer.cpp
#	src/ModelViewer.h
#	src/collider/GeomTree.cpp
#	src/collider/GeomTree.h
#	src/graphics/Material.cpp
#	src/graphics/Material.h
#	src/graphics/Types.h
#	src/graphics/VertexArray.cpp
#	src/graphics/VertexArray.h
#	src/graphics/opengl/Makefile.am
#	src/graphics/opengl/Program.cpp
#	src/graphics/opengl/RendererGL.cpp
#	src/graphics/opengl/RendererGL.h
#	src/graphics/opengl/VertexBufferGL.cpp
#	src/libs.h
#	src/scenegraph/Model.cpp
#	src/scenegraph/Model.h
#	src/vector2.h
#	src/vector3.h
#	win32/vc2013Express/pioneer.vcxproj
#	win32/vs2019/pioneer.sln
#	win32/vs2019/texturecompiler.vcxproj
#	win32/vs2019/texturecompiler.vcxproj.filters

+++ src/Asteroid.cpp
@@ -2,6 +2,7 @@
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "Asteroid.h"
+#include "graphics/RenderState.h"
 #include "graphics/Texture.h"
 #include "collider/Weld.h"
 #include "gameconsts.h"
@@ -882,7 +883,7 @@ void Asteroid::CreateAndPopulateRenderBuffers(TLOD *pLOD, Graphics::Renderer *r)
 
 	// indices
 	ib.Reset(r->CreateIndexBuffer(pLOD->m_indices.size(), BUFFER_USAGE_STATIC));
-	Uint16 *idxPtr = ib->Map(Graphics::BUFFER_MAP_WRITE);
+	Uint32 *idxPtr = ib->Map(Graphics::BUFFER_MAP_WRITE);
 	for (auto it : pLOD->m_indices) {
 		*idxPtr = it;
 		idxPtr++;
